+++ airbyte-cdk/java/airbyte-cdk/db-sources/src/main/java/io/airbyte/cdk/integrations/debezium/internals/mongodb/MongoDbCdcEventUtils.java
@@ -27,12 +27,14 @@ import org.bson.BsonReader;
 import org.bson.BsonRegularExpression;
 import org.bson.BsonType;
 import org.bson.Document;
+import org.bson.UuidRepresentation;
 import org.bson.codecs.BsonCodecProvider;
 import org.bson.codecs.BsonValueCodecProvider;
 import org.bson.codecs.DocumentCodecProvider;
 import org.bson.codecs.IterableCodecProvider;
 import org.bson.codecs.JsonObjectCodecProvider;
 import org.bson.codecs.MapCodecProvider;
+import org.bson.codecs.UuidCodecProvider;
 import org.bson.codecs.ValueCodecProvider;
 import org.bson.codecs.configuration.CodecRegistry;
 import org.bson.codecs.jsr310.Jsr310CodecProvider;
@@ -67,7 +69,12 @@ public class MongoDbCdcEventUtils {
    *         key.
    */
   public static String generateObjectIdDocument(final JsonNode debeziumEventKey) {
-    return debeziumEventKey.get(ID_FIELD).asText().replaceAll(OBJECT_ID_FIELD_PATTERN, DOCUMENT_OBJECT_ID_FIELD);
+    final String idField = debeziumEventKey.get(ID_FIELD).asText();
+    if (StringUtils.contains(idField, OBJECT_ID_FIELD)) {
+      return idField.replaceAll(OBJECT_ID_FIELD_PATTERN, DOCUMENT_OBJECT_ID_FIELD);
+    } else {
+      return Jsons.serialize(debeziumEventKey).replaceAll(ID_FIELD, DOCUMENT_OBJECT_ID_FIELD);
+    }
   }
 
   /**
@@ -213,6 +220,7 @@ public class MongoDbCdcEventUtils {
     try {
       final CodecRegistry customCodecRegistry =
           fromProviders(asList(
+              new UuidCodecProvider(UuidRepresentation.STANDARD),
               new ValueCodecProvider(),
               new BsonValueCodecProvider(),
               new DocumentCodecProvider(),