+++ airbyte-ci/connectors/pipelines/pipelines/airbyte_ci/connectors/test/commands.py
@@ -4,15 +4,15 @@
 
 import sys
 
-import anyio
-import click
+import asyncclick as click
 from pipelines import main_logger
 from pipelines.airbyte_ci.connectors.context import ConnectorContext
 from pipelines.airbyte_ci.connectors.pipeline import run_connectors_pipelines
-from pipelines.airbyte_ci.connectors.test.steps import run_connector_test_pipeline
+from pipelines.airbyte_ci.connectors.test.pipeline import run_connector_test_pipeline
 from pipelines.cli.dagger_pipeline_command import DaggerPipelineCommand
 from pipelines.consts import ContextState
 from pipelines.helpers.github import update_global_commit_status_check_for_tests
+from pipelines.helpers.utils import fail_if_missing_docker_hub_creds
 
 
 @click.command(cls=DaggerPipelineCommand, help="Test all the selected connectors.")
@@ -37,18 +37,28 @@ from pipelines.helpers.github import update_global_commit_status_check_for_tests
     type=bool,
     is_flag=True,
 )
+@click.option(
+    "--concurrent-cat",
+    help="When enabled, the CAT tests will run concurrently. Be careful about rate limits",
+    default=False,
+    type=bool,
+    is_flag=True,
+)
 @click.pass_context
-def test(
+async def test(
     ctx: click.Context,
     code_tests_only: bool,
     fail_fast: bool,
     fast_tests_only: bool,
+    concurrent_cat: bool,
 ) -> bool:
     """Runs a test pipeline for the selected connectors.
 
     Args:
         ctx (click.Context): The click context.
     """
+    if ctx.obj["is_ci"]:
+        fail_if_missing_docker_hub_creds(ctx)
     if ctx.obj["is_ci"] and ctx.obj["pull_request"] and ctx.obj["pull_request"].draft:
         main_logger.info("Skipping connectors tests for draft pull request.")
         sys.exit(0)
@@ -80,12 +90,16 @@ def test(
             fast_tests_only=fast_tests_only,
             code_tests_only=code_tests_only,
             use_local_cdk=ctx.obj.get("use_local_cdk"),
+            s3_build_cache_access_key_id=ctx.obj.get("s3_build_cache_access_key_id"),
+            s3_build_cache_secret_key=ctx.obj.get("s3_build_cache_secret_key"),
+            docker_hub_username=ctx.obj.get("docker_hub_username"),
+            docker_hub_password=ctx.obj.get("docker_hub_password"),
+            concurrent_cat=concurrent_cat,
         )
         for connector in ctx.obj["selected_connectors_with_modified_files"]
     ]
     try:
-        anyio.run(
-            run_connectors_pipelines,
+        await run_connectors_pipelines(
             [connector_context for connector_context in connectors_tests_contexts],
             run_connector_test_pipeline,
             "Test Pipeline",