+++ airbyte-integrations/connectors/source-pinterest/unit_tests/test_source.py
@@ -36,16 +36,19 @@ def test_check_wrong_date_connection(wrong_date_config):
     logger_mock = MagicMock()
     with pytest.raises(AirbyteTracedException) as e:
         source.check_connection(logger_mock, wrong_date_config)
-    assert e.value.message == "Entered `Start Date` does not match format YYYY-MM-DD"
+    assert e.value.message == "Entered `Start Date` wrong_date_format does not match format YYYY-MM-DD"
 
 
 @responses.activate
-def test_streams(test_config):
-    setup_responses()
+def test_check_connection_expired_token(test_config):
+    responses.add(responses.POST, "https://api.pinterest.com/v5/oauth/token", status=401)
     source = SourcePinterest()
-    streams = source.streams(test_config)
-    expected_streams_number = 14
-    assert len(streams) == expected_streams_number
+    logger_mock = MagicMock()
+    assert source.check_connection(logger_mock, test_config) == (
+        False,
+        "Try to re-authenticate because current refresh token is not valid. "
+        "401 Client Error: Unauthorized for url: https://api.pinterest.com/v5/oauth/token",
+    )
 
 
 def test_get_authenticator(test_config):