+++ airbyte-integrations/connectors/source-postgres/src/test/java/io/airbyte/integrations/source/postgres/CdcPostgresSourceLegacyCtidTest.java
@@ -4,11 +4,13 @@
 
 package io.airbyte.integrations.source.postgres;
 
+import org.junit.jupiter.api.Order;
+
+@Order(2)
 public class CdcPostgresSourceLegacyCtidTest extends CdcPostgresSourceTest {
 
-  @Override
-  protected String getServerImageName() {
-    return "debezium/postgres:13-alpine";
+  protected static String getServerImageName() {
+    return "debezium/postgres:13-bullseye";
   }
 
 }